Igor pathfinding opt (#39)
Yesterday: optimized the code by removing two lists this should help
with memory and my sanity.


Added My two TODOs for today.

TODO 1: Added a lineup point to create a straight path to path-find to.

This has two big advantages:

-One: Now lineups should be more consistent than a singular point this
is due to how AD* path-finding calculates his paths. Basically with a
singular point the robot has trouble rotating towards the POI but it has
less trouble with a path because now it's got two or more points of
reference.

-Two: This gives a buffer to warmup commands, activate vision-based
commands and rotating to the point saving us time.

TODO 2: Make an angle optimization method.

This optimizes the angle the robot looks toward for omnidirectional
points like the coral stands. It works by getting the orientation of
robot -> point and applying this angle to where the robot should be
looking at.